Add basic test suite
Tests are operating on reading plug-in configuration and querying
results from the SyMenu configuration file.

The file based fixtures are a 'config.json' as well as a exemplary
SyMenuItems.xml file available via SyMenuItems.zip archive.

Additionally the wox.py module is part of the testing suite. It must be
copied to plug-in root to make sure it's importable. Since Wox will
remove the wox.py from the plug-in root on every launch of the host
application, the copy is required every time Wox is launched/restarted.